parcelbee_backend/requirements.txt

Django==4.2.7
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
pillow==10.1.0
requests==2.31.0